+++ cli/balancing_services_cli/main.py
@@ -19,6 +19,7 @@ from balancing_services_cli.commands.energy import (
     energy_bids,
     energy_cbpm,
     energy_cross_border_volumes,
+    energy_day_ahead_prices,
     energy_offered,
     energy_prices,
 )
@@ -82,6 +83,7 @@ cli.add_command(energy_prices)
 cli.add_command(energy_bids)
 cli.add_command(energy_cbpm)
 cli.add_command(energy_cross_border_volumes)
+cli.add_command(energy_day_ahead_prices)
 cli.add_command(capacity_bids)
 cli.add_command(capacity_prices)
 cli.add_command(capacity_procured)